ci/supportedBranches.js: move to ci/github-script
This way it will be covered by our package.json when we configure ES Modules
in the next commit.
