+++ ci/github-script/manual-file-edits.js
@@ -1,4 +1,4 @@
-const { classify } = require('../supportedBranches.js')
+const { classify } = require('./supportedBranches.js')
 const { getCommitDetailsForPR } = require('./get-pr-commit-details')
 
 /**
