+++ ci/github-script/merge.js
@@ -1,5 +1,5 @@
 // @ts-nocheck
-const { classify } = require('../supportedBranches.js')
+const { classify } = require('./supportedBranches.js')
 
 function runChecklist({
   committers,
