+++ ci/github-script/prepare.js
@@ -1,5 +1,5 @@
 // @ts-nocheck
-const { classify } = require('../supportedBranches.js')
+const { classify } = require('./supportedBranches.js')
 const { postReview, dismissReviews } = require('./reviews.js')
 const reviewKey = 'prepare'
 const supportedSystems = require('./supportedSystems.js')
@@ -65,7 +65,7 @@ module.exports = async ({ github, context, core, dry }) => {
       // commits between that base and head is the real base. We can query for this via GitHub's
       // REST API. There can be multiple candidates for the real base with the same number of
       // commits. In this case we pick the "best" candidate by a fixed ordering of branches,
-      // as defined in ci/supportedBranches.js.
+      // as defined in ./supportedBranches.js.
       //
       // These requests take a while, when comparing against the wrong release - they need
       // to look at way more than 10k commits in that case. Thus, we try to minimize the
